Rename weatherreport.js to weatherReport.js, correct imports.
